# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lojista",
    "version": "v1"
  },
  "paths": {
    "/api/Budgets/{id}": {
      "get": {
        "tags": [
          "Budgets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Budgets/{id}/Reject": {
      "put": {
        "tags": [
          "Budgets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Budgets/{id}/Approved": {
      "put": {
        "tags": [
          "Budgets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Budget"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Solicitations": {
      "get": {
        "tags": [
          "Solicitations"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solicitation"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solicitation"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Solicitation"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Solicitations"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitation"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitation"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitation"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    },
    "/api/Solicitations/{id}": {
      "get": {
        "tags": [
          "Solicitations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Solicitations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitation"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitation"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitation"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Solicitation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Solicitations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Solicitation"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ShoppingCar": {
        "type": "object",
        "properties": {
          "codeProduct": {
            "type": "integer",
            "format": "int32"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          },
          "obs": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Solicitation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "shoppingCars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ShoppingCar"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Budget": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "dateReceive": {
            "type": "string",
            "format": "date-time"
          },
          "idSolicitation": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "boolean",
            "nullable": true
          },
          "solicitation": {
            "$ref": "#/components/schemas/Solicitation"
          }
        },
        "additionalProperties": false
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "idSolicitation": {
            "type": "integer",
            "format": "int32"
          },
          "solicitation": {
            "$ref": "#/components/schemas/Solicitation"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}